modify: separate example API usage to example project
